+++ packages/cloud-cognitive/src/components/Datagrid/utils/getArgTypes.js
@@ -75,4 +75,9 @@ export const ARG_TYPES = {
     description:
       'This value controls the height of the expanded content area. _This value is set/passed inside of the `datagridState` object._',
   },
+  customizeColumnsProps: {
+    control: 'object',
+    description:
+      'This is an object containing all of the props used with the column customization extension. _This value is set/passed inside of the `datagridState` object._',
+  },
 };