+++ packages/core/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.37.5](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive-core@0.37.4...@carbon/ibm-cloud-cognitive-core@0.37.5) (2022-10-04)
+
+**Note:** Version bump only for package @carbon/ibm-cloud-cognitive-core
+
+
+
+
+
+## [0.37.4](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive-core@0.37.3...@carbon/ibm-cloud-cognitive-core@0.37.4) (2022-09-27)
+
+
+### Bug Fixes
+
+* update to Carbon v10 compatible versions to latest ([#2276](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/2276)) ([c35a363](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/c35a3630ecb0dd60d920e57d81bd8f6e84aae008))
+
+
+
+
+
 ## [0.37.3](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive-core@0.37.2...@carbon/ibm-cloud-cognitive-core@0.37.3) (2022-09-20)
 
 **Note:** Version bump only for package @carbon/ibm-cloud-cognitive-core